Fixed the substring matching.
Substring matching was replacing * with ".*" and converting the result to a
regular expression.  This is obviously badwrong since someone could have passed
in characters that have a special meaning in the context of a regular expression.
These characters would have lead to strange behavior in the matching of property values.